requirements.txt
asgiref==3.2.8
certifi==2020.4.5.2
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.7
django-cors-headers==3.4.0
django-crum==0.7.6
django-model-utils==4.0.0
django-phone-field==1.8.1
django-post-office==3.4.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
drf-yasg==1.17.1
idna==2.9
inflection==0.5.0
itypes==1.2.0
Jinja2==2.11.2
jsonfield==3.1.0
MarkupSafe==1.1.1
packaging==20.4
PyJWT==1.7.1
pyparsing==2.4.7
pytz==2020.1
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.15.0
sqlparse==0.3.1
typing==3.7.4.1
uritemplate==3.0.1
urllib3==1.25.9
